Modify q_shuffle
In the previous version, I created a new list new_list inside the while loop, which resulted in frequent memory allocations and deallocations, impacting performance. Therefore, I modified it to create new_list outside the function and reuse it within the loop to reduce the number of memory allocations and deallocations.